+++ pellets/z_lua_bind/z_lua_bind.h
@@ -17,7 +17,6 @@
 #pragma once
 ///> 请将该文件包含到stdafx.h
 
-
 ///> 包含lua头文件，及全局定义
 #include "z_lua_bind_def.h"
 
@@ -25,7 +24,7 @@
 #include "z_lua_bind_internal.hpp"
 
 ///> 用于辅助注册一系列的函数给lua
-#include "z_lua_function_reg.h"
+#include "z_lua_function_reg.hpp"
 
 ///> 封装lua_State
-#include "z_lua_state.h"
+#include "z_lua_state.hpp"
