+++ pellets/z_lua_bind/z_lua_bind_internal.hpp
@@ -16,9 +16,9 @@
  ************************************************************************/
 #pragma once
 
-#include "z_lua_bind_internal_push.h"
-#include "z_lua_bind_internal_pop.h"
+#include "z_lua_bind_internal_push.hpp"
 #include "z_lua_bind_internal_read.hpp"
+#include "z_lua_bind_internal_pop.hpp"
 #include "z_lua_bind_internal_global.hpp"
 #include "z_lua_bind_internal_typetrait.hpp"
 #include "z_lua_bind_internal_functor.hpp"
