+++ pellets/z_lua_bind/z_lua_bind_internal_global.hpp
@@ -21,20 +21,20 @@ namespace zl
 namespace LuaBind
 {
 
-// global variable
-template<typename T>
-void set(lua_State* L, const char* name, T object)
-{
-    push(L, object);
-    lua_setglobal(L, name);
-}
+    // global variable
+    template<typename T>
+    void set(lua_State* L, const char* name, T object)
+    {
+        push(L, object);
+        lua_setglobal(L, name);
+    }
 
-template<typename T>
-T get(lua_State* L, const char* name)
-{
-    lua_getglobal(L, name);
-    return pop<T>(L);
-}
+    template<typename T>
+    T get(lua_State* L, const char* name)
+    {
+        lua_getglobal(L, name);
+        return pop<T>(L);
+    }
 
 }
 }
