+++ pellets/z_lua_bind/z_lua_bind_internal_pop.hpp
@@ -14,11 +14,25 @@
  *               Website: https://github.com/zpublic/zpublic             *
  *                                                                       *
  ************************************************************************/
-#include "stdafx.h"
-#include "z_lua_bind_internal_pop.h"
+#pragma once
 
-template<>
-void zl::LuaBind::pop(lua_State *L)
+namespace zl
 {
-    lua_pop(L, 1);
-}
+namespace LuaBind
+{
+
+    template<typename T>
+    T pop(lua_State *L)
+    {
+        T t = lua_read<T>::r(L, -1);
+        lua_pop(L, 1);
+        return t;
+    }
+
+    template<>
+    inline void pop(lua_State *L)
+    {
+        lua_pop(L, 1);
+    }
+
+}}
