+++ src/test_pellets/test_pellets.vcxproj
@@ -221,17 +221,17 @@
     <ClInclude Include="..\..\pellets\z_http\z_http_header.h" />
     <ClInclude Include="..\..\pellets\z_lua_bind\z_lua_bind.h" />
     <ClInclude Include="..\..\pellets\z_lua_bind\z_lua_bind_def.h" />
-    <ClInclude Include="..\..\pellets\z_lua_bind\z_lua_bind_internal_call.hpp" />
-    <ClInclude Include="..\..\pellets\z_lua_bind\z_lua_bind_internal_global.hpp" />
     <ClInclude Include="..\..\pellets\z_lua_bind\z_lua_bind_internal.hpp" />
-    <ClInclude Include="..\..\pellets\z_lua_bind\z_lua_bind_internal_functor.hpp" />
-    <ClInclude Include="..\..\pellets\z_lua_bind\z_lua_bind_internal_pop.h" />
-    <ClInclude Include="..\..\pellets\z_lua_bind\z_lua_bind_internal_push.h" />
+    <ClInclude Include="..\..\pellets\z_lua_bind\z_lua_bind_internal_call.hpp" />
     <ClInclude Include="..\..\pellets\z_lua_bind\z_lua_bind_internal_def.hpp" />
+    <ClInclude Include="..\..\pellets\z_lua_bind\z_lua_bind_internal_functor.hpp" />
+    <ClInclude Include="..\..\pellets\z_lua_bind\z_lua_bind_internal_global.hpp" />
+    <ClInclude Include="..\..\pellets\z_lua_bind\z_lua_bind_internal_pop.hpp" />
+    <ClInclude Include="..\..\pellets\z_lua_bind\z_lua_bind_internal_push.hpp" />
     <ClInclude Include="..\..\pellets\z_lua_bind\z_lua_bind_internal_read.hpp" />
     <ClInclude Include="..\..\pellets\z_lua_bind\z_lua_bind_internal_typetrait.hpp" />
-    <ClInclude Include="..\..\pellets\z_lua_bind\z_lua_function_reg.h" />
-    <ClInclude Include="..\..\pellets\z_lua_bind\z_lua_state.h" />
+    <ClInclude Include="..\..\pellets\z_lua_bind\z_lua_function_reg.hpp" />
+    <ClInclude Include="..\..\pellets\z_lua_bind\z_lua_state.hpp" />
     <ClInclude Include="..\..\pellets\z_net\z_net.h" />
     <ClInclude Include="..\..\pellets\z_net\z_net_socket_address.h" />
     <ClInclude Include="..\..\pellets\z_net\z_net_def.h" />
@@ -430,10 +430,6 @@
     <ClCompile Include="..\..\pellets\serialize\data_format.cpp" />
     <ClCompile Include="..\..\pellets\serialize\serialize_engine.cpp" />
     <ClCompile Include="..\..\pellets\z_http\TestHttp.cpp" />
-    <ClCompile Include="..\..\pellets\z_lua_bind\z_lua_bind_internal_pop.cpp" />
-    <ClCompile Include="..\..\pellets\z_lua_bind\z_lua_bind_internal_push.cpp" />
-    <ClCompile Include="..\..\pellets\z_lua_bind\z_lua_function_reg.cpp" />
-    <ClCompile Include="..\..\pellets\z_lua_bind\z_lua_state.cpp" />
     <ClCompile Include="..\..\pellets\z_net\z_net_socket_address_ipv4.cpp" />
     <ClCompile Include="..\..\pellets\z_thread_pool\z_task.cpp">
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">false</ExcludedFromBuild>
